Added debug & found a problem
I added a debug option to spit out an obnoxious amount of BTW data, and
at the same time found a problem where every other word was beingn
skipped, or so it seemed. So I removed the doctoring of the wordlist
from inside the loop where it's being used (which is always a terrible
idea) and instead just set self.wordlist to be the new candidate list at
the loop's end. Much better.