Improve known letter searching
Added KnownLetters class to encapsulate handling a dictionary where
letters are keys and the values are arrays of integers representing the
indices where these letters are known to _not_ be found.

Then, improved the main algorithm to not only look for known letters,
but to look for known letters in _new_ locations, and exclude those
where the letter is in a location where it cannot be.